requirements.txt
appdirs             1.4.4
asgiref             3.7.2
beautifulsoup4      4.12.2
certifi             2023.7.22
charset-normalizer  3.2.0
Django              4.2.5
django-filter       23.2
django-ninja        0.22.2
djangorestframework 3.14.0
frozendict          2.3.8
html5lib            1.1
idna                3.4
lxml                4.9.3
multitasking        0.0.11
mysqlclient         2.2.0
numpy               1.25.2
pandas              2.1.0
Pillow              10.0.0
pip                 23.2.1
pydantic            1.10.12
python-dateutil     2.8.2
pytz                2023.3.post1
requests            2.31.0
setuptools          57.4.0
six                 1.16.0
soupsieve           2.5
sqlparse            0.4.4
typing_extensions   4.7.1
tzdata              2023.3
urllib3             2.0.4
webencodings        0.5.1
yfinance            0.2.28